remove timer Registration struct
